Release/2.0.1

+++ iOSDeviceManager/Commands/iOSReturnStatusCode.h
@@ -1,4 +1,4 @@
-typedef NS_ENUM(int, iOSReturnStatusCode) {
+typedef NS_ENUM(NSUInteger, iOSReturnStatusCode) {
     iOSReturnStatusCodeEverythingOkay = 0,
     iOSReturnStatusCodeGenericFailure,
     iOSReturnStatusCodeFalse,
